changed landing page to index and routed ttmember selection to triage home
